Fix rspec tests
