+++ lib/deqm_test_kit/evaluate_measure.rb
@@ -8,21 +8,8 @@ module DEQMTestKit
   class EvaluateMeasure < Inferno::TestGroup
     # module for shared code for $evaluate-measure assertions and requests
     module MeasureEvaluationHelpers
-      def measure_evaluation_assert_success(_report_type, resource_type, params, expected_status: 200) # rubocop:disable Metrics/MethodLength
-        body = {
-          resourceType: 'Parameters',
-          parameter: [
-            {
-              name: 'periodStart',
-              valueDate: '2019-01-01'
-            },
-            {
-              name: 'periodEnd',
-              valueDate: '2019-12-31'
-            }
-          ]
-        }
-        fhir_operation("/Measure/#{measure_id}/$#{config.options[:endpoint_name]}?#{params}", body:)
+      def measure_evaluation_assert_success(_report_type, resource_type, params, expected_status: 200)
+        fhir_operation("/Measure/#{measure_id}/$#{config.options[:endpoint_name]}?#{params}")
         assert_success(resource_type, expected_status)
       end
